add threding

+++ rest_api_test.py
@@ -1,9 +1,11 @@
 import os
 import sys
+from datetime import datetime
 import urllib.request
 import base64
 import pytest
 import json
+import threading
 
 
 URL_BASE = 'http://localhost:8000/players'
@@ -13,43 +15,67 @@ class RestApi:
     """
     This class will generate basic HTTP REST API properties.
     """
-    def __init__(self, url, page_start=False, user='admin', passw='admin'):
+    def __init__(self, url, user='admin', passw='admin'):
         """
         @param::url - API url string 
-        @param::page_start - if exists
         """
+        self._url = None
         self.url_base = url
-        self.page = page_start
-        self.url = self.set_url(self.page)
         self.user_name = user
         self.password = passw
     
-    
-    def set_url(self, page):
-        """
-        The function will set url to next page
-        """
-        if not page:
-            url = self.url_base
-        else:
-            url = ('{}?page={}'.format(self.url_base, self.page))
-        
-        return url
 
+    @property
+    def url(self):
+        return self._url
+
+
+    @url.setter
+    def url(self, page):
+        if not page:
+            self._url = self.url_base
+        else:
+            self._url = ('{}?page={}'.format(self.url_base, page))
+     
+
+    @url.deleter
+    def url(self):
+        del self._url
+
+        
+
+def authorized_request(req, encoded_credentials):
+    """
+    @param::encoded_credentials - ascii format credentials
+    """
+    req.add_header('Authorization', 'Basic %s' % encoded_credentials.decode("ascii"))
+    
 
 def install_modules():
+    """
+    If modules not installed in your VM
+    """
     os.system('''wget https://bootstrap.pypa.io/get-pip.py''')
     os.system('''export PATH="$PWD/.local/bin:$PATH"''')
     os.system('''python get-pip.py''')
-    os.system('sudo easy_install pip')
+    os.system('''sudo easy_install pip''')
+    os.system('''pip install pytest==2.9.1''')
 
 
 def run_server():
-    os.system("wget -O twtask https://drive.google.com/u/0/uc?id=1V4pn_Ydu6Pzudju0tFMXqKoR0gzQyeqg&export=download")
-    os.system("chmod 755 twtask")
-    ## TODO: run server in defferent thread
-
+    """
+    run Linux executable, if needed, clone the file first.
+    """
+    if not os.path.isfile('twtask'):
+        os.system("wget -O twtask https://drive.google.com/u/0/uc?id=1V4pn_Ydu6Pzudju0tFMXqKoR0gzQyeqg&export=download")
     
+    os.system("chmod 755 twtask")
+    os.system("./twtask")
+    ## TODO: run server in defferent thread
+    print ('###### twtask #######')
+    pass
+
+
 
 def main():
     """
@@ -59,26 +85,47 @@ def main():
         out_file.write(data)
 
     """
-    Players=RestApi(URL_BASE,1)
+    t1 = threading.Thread(target=run_server, args=[])
+    t1.start()
+    page=1
+    no_error=True
+    Players=RestApi(URL_BASE)
     credentials = ('%s:%s' % (Players.user_name, Players.password))
-    encoded_credentials = base64.b64encode(credentials.encode('ascii'))
-    req = urllib.request.Request(Players.url)
-    print ('Players.url: {}'.format(Players.url))
-    req.add_header('Authorization', 'Basic %s' % encoded_credentials.decode("ascii"))
 
-    try:
-        with urllib.request.urlopen(req) as response:
-            data_bytes = response.read()
-            data_list = json.loads(data_bytes.decode("utf-8").replace("'", '"'))
-            data_json = json.dumps(data_list, indent=4, sort_keys=True)
-            print (data_json)
-           
-    except Exception as e:
-            print ('YONI: {}'.format(e))
+
+    while no_error and page<200000:
+        Players.url = page
+        req = urllib.request.Request(Players.url)
+        authorized_request(req, base64.b64encode(credentials.encode('ascii')))
+        print (Players.url)
+        del Players.url
+        try:
+            with urllib.request.urlopen(req) as response:
+                data_bytes = response.read()
+                data_list = json.loads(data_bytes.decode("utf-8").replace("'", '"'))
+                data_json = json.dumps(data_list, indent=4, sort_keys=True)
+                #print (data_json)
+                page += 1
+        
+        except Exception as e:
+                print ('ERROR handler: {}'.format(e))
+                if '111' in str(e):
+                    t1 = threading.Thread(target=run_server, args=[])
+                    t1.start()
+                    continue
+                else:
+                    no_error = False
+                
+        
+    
+    print ('Checked {} pages'.format(page))
 
 
 
 
 if __name__ == '__main__':
+    
     assert float('.'.join(map(str,sys.version_info[:2]))) > 3.6, 'python version is too old: {}'.format(sys.version_info[:2])
+    start_time = datetime.now()
     main()
+    print('Execution time: ' + str(datetime.now() - start_time) + 'ms')
